refactor(docs): separate site layout from generation
